src/requirements.txt
aiomysql==0.0.20
APScheduler==3.6.3
cffi==1.14.3
cryptography==3.2.1
numpy==1.19.4
pandas==1.1.4
pycparser==2.20
PyMySQL==0.9.2
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.20
tzlocal==2.1
